refactor(main): move each desktop ctx to separate files
