+++ main/src/platform/desktop/background/ctx.ts
@@ -1,0 +1,32 @@
+import { Config } from "@/features/config";
+import type { DesktopCtx } from "@/features/ctx";
+import { Store } from "@/features/store";
+import { LazyAsync } from "@/features/utils";
+import { DesktopAnkiApi } from "../anki";
+import { DesktopBackend } from "../backend";
+import { Database } from "../db";
+import { DesktopAnkiApiPage } from "../page/anki";
+import { DesktopPlatform } from "../platform";
+import { DesktopBackendBackground } from "./backend";
+import { DesktopPlatformBackground } from "./platform";
+
+export function createBackgroundDesktopCtx(): DesktopCtx {
+  const db = new LazyAsync(() => Database.init());
+  const platformBackground = new DesktopPlatformBackground(db);
+  const platform = DesktopPlatform.background(platformBackground);
+  const backendBackground = new DesktopBackendBackground(db);
+  const backend = DesktopBackend.background(backendBackground);
+  const store = new Store(platform);
+  const lazyConfig = new LazyAsync(() => Config.initialize(platform));
+  const ankiApiPage = new DesktopAnkiApiPage(lazyConfig);
+  const anki = DesktopAnkiApi.page(lazyConfig, ankiApiPage);
+
+  return {
+    platformType: "desktop",
+    platform,
+    store,
+    lazyConfig,
+    backend,
+    anki,
+  };
+}
