+++ main/src/platform/desktop/background/platform.ts
@@ -1,0 +1,6 @@
+import type { LazyAsync } from "@/features/utils";
+import type { Database } from "../db";
+
+export class DesktopPlatformBackground {
+  constructor(public readonly db: LazyAsync<Database>) {}
+}
