+++ main/src/platform/desktop/content/ctx.ts
@@ -1,0 +1,25 @@
+import { Config } from "@/features/config";
+import type { DesktopCtx } from "@/features/ctx";
+import { Store } from "@/features/store";
+import { LazyAsync } from "@/features/utils";
+import { DesktopAnkiApi } from "../anki";
+import { DesktopBackend } from "../backend";
+import { DesktopPlatform } from "../platform";
+
+/** Must be executed synchronously on page load */
+export function createContentDesktopCtx(): DesktopCtx {
+  const platform = DesktopPlatform.foreground();
+  const backend = DesktopBackend.foreground();
+  const store = new Store(platform);
+  const lazyConfig = new LazyAsync(() => Config.initialize(platform));
+  const anki = DesktopAnkiApi.content(lazyConfig);
+
+  return {
+    platformType: "desktop",
+    platform,
+    store,
+    lazyConfig,
+    backend,
+    anki,
+  };
+}
