+++ main/src/platform/desktop/index.ts
@@ -1,4 +1,11 @@
 export { DesktopAnkiApi } from "./anki";
 export { DesktopBackend } from "./backend";
-export { createBackgroundDesktopCtx, createContentDesktopCtx, createPageDesktopCtx } from "./ctx";
 export { DesktopPlatform } from "./platform";
+
+/*
+Do not export here because these run global scripts and thus not tree shaked if exported.
+
+export { createBackgroundDesktopCtx } from "./background/ctx";
+export { createContentDesktopCtx } from "./content/ctx";
+export { createPageDesktopCtx } from "./page/ctx";
+*/
